spark/requirements.txt
faker
pandas
numpy
matplotlib
geopandas
pyarrow
shapely
folium
osmnx
psycopg2
apache-airflow
